refactor(promql): Actually rename promflux package to promql
