+++ promql/query_helpers.go
@@ -1,4 +1,4 @@
-package promflux
+package promql
 
 import (
 	"sort"
