+++ promql/query_helpers_test.go
@@ -1,4 +1,4 @@
-package promflux
+package promql
 
 import (
 	"bytes"
